Fix a variable overriding/redefinition bug
A variable named $l is used for working purpose in an if statement, but
this destroys the previously defined variable with the same name that
holds a source line number because they are in the same scope. A
subsequent use of the line number variable causes a crash.

The solution implemented here is to rename the work variable.